+++ Synthie/Flange.h
@@ -26,7 +26,7 @@ private:
 	double m_wet;
 	double m_delay;
 	
-	std::vector<short> m_queue;
+	std::vector<double> m_queue;
 	int m_wrloc;
 	int m_rdloc;
 };
